refactor: extract duplicated code getClientIp
* getting the client ip was duplicated in cap.js and ratelimit.js which
  meant that if it needs changes in the future we would have to change
two functions that effectively do the same
* introduced new file ip_extraction that exports the function
  getClientIp, the function is then imported in cap.js and ratelimit.js
* dropped one warning in ratelimiter that the IP could not be found in
  the header because it is duplicated, as we log "Unable to determine
client IP" anyways if it cannot be found
